+++ MaxMind.MinFraud.UnitTest/WebServiceClientTest.cs
@@ -53,6 +53,9 @@ namespace MaxMind.MinFraud.UnitTest
             Assert.True(response.IPAddress.RegisteredCountry.IsInEuropeanUnion);
             Assert.True(response.IPAddress.RepresentedCountry.IsInEuropeanUnion);
             Assert.True(response.IPAddress.Traits.IsResidentialProxy);
+
+            Assert.Equal("310", response.IPAddress.Traits.MobileCountryCode);
+            Assert.Equal("004", response.IPAddress.Traits.MobileNetworkCode);
         }
 
         [Fact]
@@ -135,7 +138,7 @@ namespace MaxMind.MinFraud.UnitTest
         }
 #endif
 
-        private void CompareJson(string responseContent, object response )
+        private void CompareJson(string responseContent, object response)
         {
             var expectedResponse = JsonDocument.Parse(responseContent);