+++ releasenotes.md
@@ -4,6 +4,11 @@ Release Notes
 3.3.0
 ------------------
 
+* Upgraded `MaxMind.GeoIP2` to 4.1.0. This adds mobile country code (MCC)
+  and mobile network code (MNC) to minFraud Insights and Factors responses.
+  These are available at `response.IPAddress.Traits.MobileCountryCode` and
+  `response.IPAddress.Traits.MobileNetworkCode`. We expect this data to be
+  available by late January 2022.
 * Added the following new values to the `PaymentProcessor` enum:
   * `Boacompra`
   * `Boku`